pyproject.toml
[project]
name = "docagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.3",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "matplotlib>=3.10.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-pptx>=1.0.2",
    "seaborn>=0.13.2",
    "spire-doc>=13.6.4",
    "spire-pdf>=11.6.1",
    "spire-presentation>=10.1.0",
    "spire-xls>=15.5.0",
]
